+++ src/runtime/crt/memory/stack_allocator.c
@@ -18,16 +18,19 @@
  */
 // LINT_C_FILE
 #include <tvm/runtime/crt/stack_allocator.h>
-#ifdef TVM_CRT_DEBUG
+#if TVM_CRT_STACK_ALLOCATOR_ENABLE_FIFO_CHECK > 0
 #include <tvm/runtime/crt/logging.h>
 #endif
 
 void* StackMemoryManager_Allocate(tvm_workspace_t* tvm_runtime_workspace, int32_t nbytes) {
-  uint32_t offset_bytes = (~nbytes + 1) & (TVM_RUNTIME_ALLOC_ALIGNMENT_BYTES - 1);
+  // reserve bytes at the end of the allocation such that
+  // next_alloc % TVM_RUNTIME_ALLOC_ALIGNMENT_BYTES == 0.
+  uint32_t offset_bytes =
+      (TVM_RUNTIME_ALLOC_ALIGNMENT_BYTES - nbytes) & (TVM_RUNTIME_ALLOC_ALIGNMENT_BYTES - 1);
   uint8_t* current_alloc = tvm_runtime_workspace->next_alloc;
   uint8_t* next_alloc = tvm_runtime_workspace->next_alloc + nbytes + offset_bytes;
   uint8_t* workspace_end = tvm_runtime_workspace->workspace + tvm_runtime_workspace->workspace_size;
-#ifdef TVM_CRT_DEBUG
+#if TVM_CRT_STACK_ALLOCATOR_ENABLE_FIFO_CHECK > 0
   const uint32_t total_size = (nbytes + offset_bytes + STACK_ALLOCATOR_TAG_SIZE_BYTES);
   *((uint32_t*)next_alloc) = total_size ^ STACK_ALLOCATOR_TAG;
   next_alloc += STACK_ALLOCATOR_TAG_SIZE_BYTES;
@@ -41,10 +44,11 @@ void* StackMemoryManager_Allocate(tvm_workspace_t* tvm_runtime_workspace, int32_
 }
 
 tvm_crt_error_t StackMemoryManager_Free(tvm_workspace_t* tvm_runtime_workspace, void* ptr) {
-#ifdef TVM_CRT_DEBUG
+#if TVM_CRT_STACK_ALLOCATOR_ENABLE_FIFO_CHECK > 0
   uint32_t tag = *(((uint32_t*)tvm_runtime_workspace->next_alloc) - 1);
-  uint32_t total_size = (tvm_runtime_workspace->next_alloc - (uint8_t*)ptr);
-  CHECK_EQ(tag, total_size ^ STACK_ALLOCATOR_TAG, "tag did not match");
+  uint32_t actual_size = (tvm_runtime_workspace->next_alloc - (uint8_t*)ptr);
+  uint32_t expected_size = tag ^ STACK_ALLOCATOR_TAG;
+  CHECK_EQ(expected_size, actual_size, "Deallocation not in FIFO ordering");
 #endif
   tvm_runtime_workspace->next_alloc = ptr;
   return 0;
