+++ tests/cpp/utvm_runtime_standalone_test.cc
@@ -37,7 +37,7 @@
 #include <tvm/relay/expr.h>
 #include <tvm/relay/transform.h>
 #include <tvm/relay/type.h>
-#include <tvm/runtime/meta_data.h>
+#include <tvm/runtime/executor_info.h>
 #include <tvm/runtime/micro/standalone/utvm_runtime.h>
 #include <tvm/runtime/module.h>
 #include <tvm/runtime/packed_func.h>
