+++ tests/python/relay/test_backend_graph_executor.py
@@ -150,10 +150,12 @@ def test_plan_memory():
     assert len(storage_sizes) == 4
 
     # Check the specific size of each sid
-    assert storage_sizes[0][0] == 40
-    assert storage_sizes[1][0] == 4
-    assert storage_sizes[2][0] == 4
-    assert storage_sizes[3][0] == 40
+    assert (
+        storage_sizes[0][0] == 40
+        and storage_sizes[1][0] == 4
+        and storage_sizes[2][0] == 4
+        and storage_sizes[3][0] == 40
+    )
 
 
 def test_reshape_nop():
